requirements.txt
google_trans_new==1.1.9
langid==1.1.6
Pillow==10.0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
transformers==4.24.0
